Moved service files into an explicit service folder
