+++ src/app/utils/services/arcane-symbol.service.ts
