+++ src/app/utils/services/reset-timer.service.ts
